+++ ignite/cmd/chain_build.go
@@ -101,7 +101,10 @@ func chainBuildHandler(cmd *cobra.Command, _ []string) error {
 		releaseTargets, _ = cmd.Flags().GetStringSlice(flagReleaseTargets)
 		releasePrefix, _  = cmd.Flags().GetString(flagReleasePrefix)
 		output, _         = cmd.Flags().GetString(flagOutput)
-		session           = cliui.New(cliui.WithVerbosity(getVerbosity(cmd)), cliui.StartSpinner())
+		session           = cliui.New(
+			cliui.WithVerbosity(getVerbosity(cmd)),
+			cliui.StartSpinner(),
+		)
 	)
 
 	defer session.End()
@@ -120,7 +123,7 @@ func chainBuildHandler(cmd *cobra.Command, _ []string) error {
 		chainOption = append(chainOption, chain.CheckDependencies())
 	}
 
-	c, err := newChainWithHomeFlags(cmd, chainOption...)
+	c, err := NewChainWithHomeFlags(cmd, chainOption...)
 	if err != nil {
 		return err
 	}