+++ ignite/cmd/scaffold_package.go
@@ -45,10 +45,8 @@ func createPacketHandler(cmd *cobra.Command, args []string) error {
 		appPath      = flagGetPath(cmd)
 	)
 
-	session := cliui.New(cliui.StartSpinner())
+	session := cliui.New(cliui.StartSpinnerWithText(statusScaffolding))
 	defer session.End()
-
-	session.StartSpinner("Scaffolding...")
 
 	module, err := cmd.Flags().GetString(flagModule)
 	if err != nil {